backend/requirements/base.txt
Django>=3.2.9,<3.3
django-cors-headers==3.10.0
djangorestframework>=3.12.4,<3.13
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
